finished implementing freight/freight.js module
